+++ src/LibraryManager/Json/LibraryStateToFileConverter.cs
@@ -6,6 +6,7 @@ using System.Diagnostics.CodeAnalysis;
 using System.Text;
 using Microsoft.Web.LibraryManager.Contracts;
 using Microsoft.Web.LibraryManager.LibraryNaming;
+using Microsoft.Web.LibraryManager.Utilities;
 
 namespace Microsoft.Web.LibraryManager.Json
 {
@@ -30,7 +31,7 @@ namespace Microsoft.Web.LibraryManager.Json
             string provider = string.IsNullOrEmpty(stateOnDisk.ProviderId) ? _defaultProvider : stateOnDisk.ProviderId;
 
             (string name, string version) = LibraryIdToNameAndVersionConverter.Instance.GetLibraryNameAndVersion(stateOnDisk.LibraryId, provider);
-            string destination = string.IsNullOrEmpty(stateOnDisk.DestinationPath) ? ExpandDestination(_defaultDestination, name, version) : stateOnDisk.DestinationPath;
+            string destination = string.IsNullOrEmpty(stateOnDisk.DestinationPath) ? PathTemplateUtility.ExpandPathTemplate(_defaultDestination, name, version) : stateOnDisk.DestinationPath;
 
             var state = new LibraryInstallationState()
             {
@@ -45,31 +46,6 @@ namespace Microsoft.Web.LibraryManager.Json
             };
 
             return state;
-        }
-
-        /// <summary>
-        /// Expands [Name] and [Version] tokens in the DefaultDestination
-        /// </summary>
-        /// <param name="destination">The default destination string</param>
-        /// <param name="name">Package name</param>
-        /// <param name="version">Package version</param>
-        /// <returns></returns>
-        [SuppressMessage("Globalization", "CA1307:Specify StringComparison for clarity", Justification = "Not available on net481, not needed here (caseless)")]
-        private string ExpandDestination(string destination, string name, string version)
-        {
-            if (destination is null || !destination.Contains("["))
-            {
-                return destination;
-            }
-
-            // if the name contains a slash (either filesystem or scoped packages),
-            // trim that and only take the last segment.
-            int cutIndex = name.LastIndexOfAny(['/', '\\']);
-
-            StringBuilder stringBuilder = new StringBuilder(destination);
-            stringBuilder.Replace("[Name]", cutIndex == -1 ? name : name.Substring(cutIndex + 1));
-            stringBuilder.Replace("[Version]", version);
-            return stringBuilder.ToString();
         }
 
         public LibraryInstallationStateOnDisk ConvertToLibraryInstallationStateOnDisk(ILibraryInstallationState state)